+++ rhine-gloss/src/FRP/Rhine/Gloss/Pure.hs
@@ -30,11 +30,14 @@ import Control.Monad.Trans.Class
 import Control.Monad.Trans.Reader
 import Control.Monad.Trans.Writer.Strict
 
--- monad-schedule
-import Control.Monad.Schedule.Class
-import Control.Monad.Schedule.Yield
-
 -- automaton
+
+import Data.Automaton.Schedule (
+  MonadSchedule (..),
+  SkipT,
+  runSkipT,
+  skip,
+ )
 import Data.Automaton.Trans.Except (performOnFirstSample)
 import qualified Data.Automaton.Trans.Reader as AutomatonReader
 import qualified Data.Automaton.Trans.Writer as AutomatonWriter
@@ -48,12 +51,12 @@ import FRP.Rhine.Gloss.Common
 -- * @gloss@ effects
 
 -- | A pure monad in which all effects caused by the @gloss@ backend take place.
-newtype GlossM a = GlossM {unGlossM :: YieldT (ReaderT (Float, Maybe Event) (Writer Picture)) a}
+newtype GlossM a = GlossM {unGlossM :: SkipT (ReaderT (Float, Maybe Event) (Writer Picture)) a}
   deriving (Functor, Applicative, Monad)
 
 -- Would have liked to make this a derived instance, but for some reason deriving gets thrown off by the newtype
 instance MonadSchedule GlossM where
-  schedule actions = fmap (fmap (fmap GlossM)) $ GlossM $ schedule $ fmap unGlossM actions
+  schedule = fmap (hoistS unGlossM) >>> schedule >>> hoistS GlossM
 
 -- | Add a picture to the canvas.
 paint :: Picture -> GlossM ()
@@ -82,7 +85,7 @@ instance Semigroup GlossClock where
 instance Clock GlossM GlossClock where
   type Time GlossClock = Float
   type Tag GlossClock = Maybe Event
-  initClock _ = return (constM (GlossM $ yield >> lift ask) >>> (sumS *** Category.id), 0)
+  initClock _ = pure (constM (GlossM (skip >> lift ask)) >>> (sumS *** Category.id), 0)
   {-# INLINE initClock #-}
 
 instance GetClockProxy GlossClock
@@ -126,7 +129,7 @@ flowGloss GlossSettings {..} rhine =
   play display backgroundColor stepsPerSecond (worldAutomaton, Blank) getPic handleEvent simStep
   where
     worldAutomaton :: WorldAutomaton
-    worldAutomaton = AutomatonWriter.runWriterS $ AutomatonReader.runReaderS $ hoistS (runYieldT . unGlossM) $ performOnFirstSample $ eraseClock rhine
+    worldAutomaton = AutomatonWriter.runWriterS $ AutomatonReader.runReaderS $ hoistS (runSkipT . unGlossM) $ performOnFirstSample $ eraseClock rhine
     stepWith :: (Float, Maybe Event) -> (WorldAutomaton, Picture) -> (WorldAutomaton, Picture)
     stepWith (diff, eventMaybe) (automaton, _) = let Result automaton' (picture, _) = runIdentity $ stepAutomaton automaton ((diff, eventMaybe), ()) in (automaton', picture)
     getPic (_, pic) = pic