+++ rhine/src/FRP/Rhine.hs
@@ -45,7 +45,6 @@ import FRP.Rhine.Clock.Realtime.Never as X
 import FRP.Rhine.Clock.Realtime.Stdin as X
 import FRP.Rhine.Clock.Select as X
 import FRP.Rhine.Clock.Trivial as X
-import FRP.Rhine.Clock.Unschedule as X
 
 import FRP.Rhine.ResamplingBuffer.ClSF as X
 import FRP.Rhine.ResamplingBuffer.Collect as X