+++ rhine/src/FRP/Rhine/Schedule.hs
@@ -8,7 +8,7 @@
 {-# LANGUAGE TypeFamilies #-}
 
 {- |
-The 'MonadSchedule' class from the @monad-schedule@ package is the compatibility mechanism between two different clocks.
+The 'MonadSchedule' class is the compatibility mechanism between two different clocks.
 It implements a concurrency abstraction that allows the clocks to run at the same time, independently.
 Several such clocks running together form composite clocks, such as 'ParallelClock' and 'SequentialClock'.
 This module defines these composite clocks,
@@ -20,37 +20,21 @@ module FRP.Rhine.Schedule where
 import Control.Arrow
 import Data.List.NonEmpty as N
 
--- monad-schedule
-import Control.Monad.Schedule.Class
-
 -- automaton
 import Data.Automaton hiding (toStreamT)
-import Data.Stream.Optimized (OptimizedStreamT (..), toStreamT)
+import Data.Automaton.Schedule
 
 -- rhine
 import FRP.Rhine.Clock
-import FRP.Rhine.Schedule.Internal
 
 -- * Scheduling
-
-{- | Run several automata concurrently.
-
-Whenever one automaton outputs a value,
-it is returned together with all other values that happen to be output at the same time.
--}
-scheduleList :: (Monad m, MonadSchedule m) => NonEmpty (Automaton m a b) -> Automaton m a (NonEmpty b)
-scheduleList automatons0 =
-  Automaton $
-    Stateful $
-      scheduleStreams' $
-        toStreamT . getAutomaton <$> automatons0
 
 {- | Run two automata concurrently.
 
 Whenever one automaton returns a value, it is returned.
 -}
 schedulePair :: (Monad m, MonadSchedule m) => Automaton m a b -> Automaton m a b -> Automaton m a b
-schedulePair automatonL automatonR = concatS $ fmap toList $ scheduleList $ automatonL :| [automatonR]
+schedulePair automatonL automatonR = schedule $ automatonL :| [automatonR]
 
 -- | Run two running clocks concurrently.
 runningSchedule ::